# docs/v150-openapi.yaml
openapi: 3.0.3
info:
  title: 7zi API
  description: |
    # 7zi - AI 驱动的团队管理平台 API
    
    11 位 AI 成员 · 24/7 自主工作 · 实时协作
    
    ## 功能特性
    - 🤖 AI Agent 智能调度系统
    - 🔄 WebSocket 实时协作
    - 🔐 RBAC 权限控制
    - 📊 性能监控与告警
    - 🌍 多语言国际化
    
    ## 认证方式
    所有 API 请求需要在 Header 中携带 JWT Token:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## 错误响应格式
    所有错误响应遵循统一格式:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Error description",
        "details": {}
      }
    }
    ```
    
    ## 限流策略
    - 默认限制: 100 请求/分钟/IP
    - 认证用户: 200 请求/分钟/用户
    - API Key 用户: 500 请求/分钟
    
  version: 1.5.0
  contact:
    name: 7zi API Support
    email: api@7zi.com
    url: https://7zi.com/docs/api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://7zi.com/api
    description: 生产环境
  - url: http://localhost:3000/api
    description: 开发环境
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 认证令牌
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: 请求的资源不存在
...
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          default: true
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          default: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - VALIDATION_ERROR
                - INTERNAL_ERROR
                - RATE_LIMIT_EXCEEDED
                - DATABASE_ERROR
                - CONFLICT
            message:
              type: string
            details:
              type: object
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          description: 总记录数
        page:
          type: integer
          description: 当前页码
        limit:
          type: integer
          description: 每页数量
        totalPages:
          type: integer
          description: 总页数

    # ===== 认证 Schema =====
...
    Feedback:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [bug, feature, improvement, other]
        title:
          type: string
        description:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        status:
          type: string
          enum: [pending, reviewed, resolved]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        userId:
          type: string
        email:
          type: string
        images:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              size:
                type: integer
              type:
                type: string
        adminNotes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
    FeedbackResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Feedback'
    CreateFeedbackRequest:
      type: object
      required: [type, title, rating]
      properties:
        type:
          type: string
          enum: [bug, feature, improvement, other]
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        rating:
          type: integer
          minimum: 1
          maximum: 5
        email:
          type: string
          format: email
        images:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              size:
                type: integer
              type:
                type: string
    FeedbackListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            feedbacks:
              type: array
              items:
                $ref: '#/components/schemas/Feedback'
            meta:
              $ref: '#/components/schemas/PaginationMeta'
            stats:
              type: object
              properties:
                total:
                  type: integer
                average:
                  type: number
                byType:
                  type: object
                  additionalProperties:
                    type: integer
                byStatus:
                  type: object
                  additionalProperties:
                    type: integer

    # ===== 健康检查 Schema =====
...
paths:
...
  /feedback:
    get:
      tags: [Feedback]
      summary: 获取反馈列表
      operationId: listFeedback
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [bug, feature, improvement, other]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, reviewed, resolved]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackListResponse'

    post:
      tags: [Feedback]
      summary: 提交反馈
      operationId: createFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeedbackRequest'
      responses:
        '201':
          description: 提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
  /feedback/{feedbackId}:
    get:
      tags: [Feedback]
      summary: 获取反馈详情
      operationId: getFeedback
      security:
        - bearerAuth: []
      parameters:
        - name: feedbackId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Feedback]
      summary: 更新反馈 (管理员)
      operationId: updateFeedback
      security:
        - bearerAuth: []
      parameters:
        - name: feedbackId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, reviewed, resolved]
                priority:
                  type: string
                  enum: [low, medium, high, urgent]
                adminNotes:
                  type: string
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'

    delete:
      tags: [Feedback]
      summary: 删除反馈 (管理员)
      operationId: deleteFeedback
      security:
        - bearerAuth: []
      parameters:
        - name: feedbackId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================================================
  # 健康检查 API
  # ============================================================================
...